updated vm and entities

+++ DataModel/Entities/UserRole.cs
@@ -11,8 +11,8 @@ namespace DataModel.Entities
     public class UserRole
     {
         [Key]
-        public int UserRoleId { get; set; }
-        public int UserId { get; set; }
-        public int RoleId { get; set; }
+        public long UserRoleId { get; set; }
+        public long UserId { get; set; }
+        public long RoleId { get; set; }
     }
 }
